Fix loading hash sets, update linux release. Add shorter synonym --use-hash-sets for --use-full-khash-sets
