--- src/sketchcmpcrmh.cpp
@@ -1,1 +1,0 @@
-#include "sketch_and_cmp.h"
